Typescript, R.I.P.
