+++ .gitignore
@@ -1,7 +1,6 @@
 # babel dist
 /lib
 /packages/*/lib
-/packages/skema/src/*.js
 
 # Numerous always-ignore extensions
 *.bak
