+++ packages/skema/package.json
@@ -4,7 +4,9 @@
   "description": "Skema provides a handy and composable way to validate and transform JavaScript objects.",
   "main": "index.js",
   "scripts": {
-    "build": "../../node_modules/.bin/babel src --out-dir lib",
+    "build": "npm run build-ts && npm run build-js",
+    "build-ts": "../../node_modules/.bin/tsc",
+    "build-js": "../../node_modules/.bin/babel src --out-dir lib",
     "test": "../../node_modules/.bin/nyc ../../node_modules/.bin/ava --verbose --timeout=10s",
     "test-no-cov": "../../node_modules/.bin/ava --verbose --timeout=10s",
     "prepublish": "npm run build",
@@ -40,7 +42,7 @@
   "dependencies": {
     "err-object": "^2.0.3",
     "make-array": "^1.0.1",
-    "promise-faker": "^1.0.3",
+    "promise-faker": "^1.2.0",
     "promise.extra": "^2.1.0",
     "symbol-for": "^1.0.0"
   }
